# Validate config using jsonschema