- [ ] Tests added / passed
- [x] Passes `black dask` / `flake8 dask`
- [x] Closes #5695

I thought I'd open this to start the discussion around validation. I am not sure that this is the right approach. This is probably too aggressive since it'll make `import dask` fail if there is a bad config file. But maybe that *is* the right behavior? Obviously we can choose a much more user-facing error. 

![Screenshot from 2020-07-24 17-13-10](https://user-images.githubusercontent.com/4806877/88435804-ff662f80-cdd0-11ea-9dca-4e02400ecbfa.png)

Here are some timings:

```python 

# with jsonschema import
>>> %timeit from dask.config import config
616 ns ± 6.08 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

# without
>>> %timeit from dask.config import config
622 ns ± 17.1 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)