@gforsyth 2020-07-27
I'm +1 on failing for broken configs, or very loud warnings (but my preference is a failure), to let me know I screwed something up.

I'm trying to think of a case where an old config would cause an unwanted failure here, but if someone is updating, wouldn't they want a nudge that they need to update the config, too?

I suppose there could be problems with users with several dask environments on a local with a shared config in `$XDG_DATA_DIR`?  I don't have a sense of how stable the full set of configuration options has been.