@jsignell 2020-07-27
To be clear this only fails on keys that are defined in the schema. So it is fine to have spare keys that the schema doesn't know about. I will make a prettier error and then maybe we can get a better sense of how we feel about it?