@mrocklin 2020-08-07
```python
# with jsonschema import
>>> %timeit from dask.config import config
616 ns ± 6.08 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

# without
>>> %timeit from dask.config import config
622 ns ± 17.1 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
```

Note that Python caches imports, so you'll probably want to use just `time`.  Also, can I ask for the same thing with importing `dask.distributed`?