@jsignell 2020-08-07
> Note that Python caches imports, so you'll probably want to use just `time`. 

Thanks! That does make sense...

> Also, can I ask for the same thing with importing `dask.distributed`?

I'm trying to figure out how to have a global list of validators that other submodules can append to. Is that what you were thinking?