@mrocklin 2020-08-07
I just meant test import speed.  distributed's config is an order of magnitude larger than dask's.  Assuming that we're also validating that (which may not be the case, I haven't gotten into this PR yet) I would recommend benchmarking any slowdown's against that config .